% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.R
\name{boot_dists}
\alias{boot_dists}
\title{Simple bootstrap for distances}
\usage{
boot_dists(df, n = 1000, B = 1000)
}
\arguments{
\item{df}{A data.frame from assembling differences}
}
\value{
A vector of bootstrap replicate means
}
\description{
Simple bootstrap for distances
}

