% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_distances.R
\name{calc_dist_ratio}
\alias{calc_dist_ratio}
\title{Calculate the distance (OSA) ratio from distance and max distance matrices}
\usage{
calc_dist_ratio(dists, maxes)
}
\arguments{
\item{dists}{A matrix of string edit distances (OSA)}

\item{maxes}{A matrix of maximum possible string edit distances}
}
\value{
A matrix of edit distance ratios
}
\description{
Raw string edit distances are of limited use because differences in the 
lengths of compared strings drives the distribution of expected distances.
The raw edit distances are normalized by element-wise division of the OSA
distance by the maximum possible distance (sum of compared string lengths).
Identical strings have a ratio of 0, completely different strings == 1.
}
\seealso{
\link{get_max_dist}
}

