% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_distances.R
\name{create_mins_df}
\alias{create_mins_df}
\title{Create a data.frame of minimum ratio string matches}
\usage{
create_mins_df(dist_list, name)
}
\arguments{
\item{name}{The name of the current (single) chapter being compared to all}

\item{df}{A list created in \link{one_vs_all_dists}}
}
\value{
A data.frame with best match data
}
\description{
To analyze and plot sentence similarity, we need the data out of the 
(ridiculous) nested list and into a useful data.frame, which this function
provides.
}
\seealso{
\link{one_vs_all_dists}
}

