% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_distances.R
\name{get_max_dist}
\alias{get_max_dist}
\title{Calculate maximum possible edit distance between x, y}
\usage{
get_max_dist(x, y)
}
\arguments{
\item{x}{A vector of strings}

\item{y}{A second vector of strings}
}
\value{
A matrix of maximum edit distances for each combination of strings
in x, y
}
\description{
The maximum edit distance is needed to calculate the distance (e.g.,
Levenshtein) ratio.
}
\examples{
\dontrun{
get_max_dist(x = sents1, y = sents2)
}
}

