% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_distances.R
\name{get_min_dists}
\alias{get_min_dists}
\title{Get the minimum edit distance for rows in a matrix}
\usage{
get_min_dists(mat)
}
\arguments{
\item{mat}{A matrix of edit distances}
}
\value{
The minimum edit distance for each row of matrix mat
}
\description{
Applys \link{get_min_dist} over all rows of matrix \code{mat}
}
\examples{
\dontrun{
get_min_dist(mat=amat, x=2)
}
}

