% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_distances.R
\name{one_vs_all_dists}
\alias{one_vs_all_dists}
\title{Calculate minimum string distances for one chapter vs. all other chapters}
\usage{
one_vs_all_dists(all_ls, one_ch, ch_name)
}
\arguments{
\item{all_ls}{A list of all chapter sentences}
}
\value{
A data.frame with minimum distances for each sentence in one_ch
}
\description{
More detailed description of the function
}
\examples{
\dontrun{
ch1_dists <- one_vs_all_dists(hcp_cur, hcp_rev)
}
}
\seealso{
if any see alsos
}

