% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_distances.R
\name{summary_dists}
\alias{summary_dists}
\title{Summarize a vector of edit distances}
\usage{
summary_dists(x)
}
\arguments{
\item{x}{A vector of edit distances}
}
\value{
A list with N, mean, median, and sd.
}
\description{
Summarize a vector of edit distances
}
\examples{
\dontrun{
summary_dists(min_dists)
}
}
\seealso{
if any see alsos
}

